+++ src/components/guest/GuestBookCard/GuestBookCard.tsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 import * as GuestBookCardStyles from "./GuestBookCard.style";
 import styled from "@emotion/styled";
 
@@ -7,23 +9,22 @@ export interface GuestBookCardProps {
     content: string;
 }
 
-export const GuestBookCard = ({ to, from, content }: GuestBookCardProps) => {
+export const GuestBookCard = React.forwardRef<HTMLDivElement, GuestBookCardProps>(({ to, from, content }, ref) => {
     return (
-        <GuestBookCardStyles.Container>
+        <GuestBookCardStyles.Container ref={ref}>
             <CardTo>TO. {to}</CardTo>
             <CardContent>{content}</CardContent>
             <CardFrom>FROM. {from}</CardFrom>
         </GuestBookCardStyles.Container>
     );
-};
-
+});
 const CardTo = styled.p`
     display: flex;
     align-items: center;
     height: 30px;
 `;
 const CardContent = styled.p`
-    height: 120px;
+    max-height: 120px;
 `;
 const CardFrom = styled.p`
     display: flex;
